server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Gemsmith <nobuhiro_ito@nii.ac.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^3.12.1"
python-dotenv = "^0.19.1"
Flask = "^2.0.2"
line-bot-sdk = "^2.0.1"
waitress = "^2.0.0"
firebase-admin = "^5.1.0"
google-cloud-firestore = "^2.3.4"
Flask-Cors = "^3.0.10"
numpy = "^1.22.3"
google-cloud-vision = "^2.7.2"
opencv-python = "^4.5.5"
imageio = "^2.16.1"
google-cloud = "^0.34.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
